+++ coretech/vision/robot/src/detectFiducialMarkers.cpp
@@ -35,7 +35,7 @@ namespace Anki
       const f32 component_minHollowRatio,
       const s32 quads_minQuadArea, const s32 quads_quadSymmetryThreshold, const s32 quads_minDistanceFromImageEdge,
       const f32 decode_minContrastRatio,
-      const s32 maxConnectedComponentSegments,
+      const u16 maxConnectedComponentSegments,
       const s32 maxExtractedQuads,
       const s32 quadRefinementIterations,
       const bool returnInvalidMarkers,
